+++ apps/pi-das/src/routers/participants.ts
@@ -19,9 +19,9 @@
 
 import { Router } from 'express';
 
-import { getParticipant, getParticipants } from '../service/search.js';
-import { createParticipant } from '../service/create.js';
-import { updateParticipant } from '../service/update.js';
+import { getParticipant, getParticipants } from '../services/search.js';
+import { createParticipant } from '../services/create.js';
+import { updateParticipant } from '../services/update.js';
 import logger from '../logger.js';
 
 // TODO: update JSDoc comments when custom error handling is implemented
